Iceman stuff
1. Added a death animation for the iceman. 2. Changed the hammer works
to react to ice pieces as armour. Ice breaks on hit before hitpoints are
reduced. 3. Fixed a bug with the water. It was reacting to contact with
the hammer. Now it only reacts to the player. 4. Added the Iceman to the
Floe overworld. They now work properly and have personalized
scripts(Same as dog basically). 5. Added a fragment particle system to
run on breaking ice armour, but armour is currently destroyed before it
runs. A problem for when we do touch ups.